AI_Agent_Basic/requirements.txt
python-telegram-bot==21.10
openai==1.75.0
python-dotenv==1.0.1
